Accordions

Accordion collapsible content. panel is a helper function to facilitate building accordion panels, see details for more information.

accordion(...)

panel(heading, body)

Arguments

...
panels converted into accordion panels or named arguments passed as HTML attributes to the accordion parent element. Note, a panel or custom panel named show will render in an open state rather than the default collapsed state.

heading
A panel heading, character string or tag element(s).

body
Body content for the panel, character string or tag element(s).

Details

panel is little more than a thin wrapper around list to ensure each accordion panel has a heading and a body. If you want more customization over accordion panels you may ignore panel and pass unnamed tag elements. When creating custom panels be sure to follow the data-target, data-toggle conventions, for more on how to build custom accordions check out the Bootstrap collapse reference. accordion will generate an HTML id if one is not specified and child panels are assigned IDs based on the id of the accordion parent.

Examples

accordion(
  panel(
    "Heading 1",
    "Body text"
  ),
  panel(
    "Heading 2",
    "Body text"
  ),
  panel(
    "Heading 3",
    "Body text"
  )
)

#> <div role="tablist" aria-multiselectable="true" id="accordion-81-835">
#>   <div class="card">
#>     <div class="card-header" role="tab">
#>       <h5 class="mb-0">
#>         <a data-toggle="collapse" data-parent="#accordion-81-835" href="#accordion-81-835-1">Heading 1</a>
#>       </h5>
#>     </div>
#>     <div id="accordion-81-835-1" class="collapse" role="tabpanel">
#>       <div class="card-block">Body text</div>
#>     </div>
#>   </div>
#>   <div class="card">
#>     <div class="card-header" role="tab">
#>       <h5 class="mb-0">
#>         <a data-toggle="collapse" data-parent="#accordion-81-835" href="#accordion-81-835-2">Heading 2</a>
#>       </h5>
#>     </div>
#>     <div id="accordion-81-835-2" class="collapse" role="tabpanel">
#>       <div class="card-block">Body text</div>
#>     </div>
#>   </div>
#>   <div class="card">
#>     <div class="card-header" role="tab">
#>       <h5 class="mb-0">
#>         <a data-toggle="collapse" data-parent="#accordion-81-835" href="#accordion-81-835-3">Heading 3</a>
#>       </h5>
#>     </div>
#>     <div id="accordion-81-835-3" class="collapse" role="tabpanel">
#>       <div class="card-block">Body text</div>
#>     </div>
#>   </div>
#> </div>