Badges, tags

A badge is a useful means of highlighting text.

badge(content = NULL, context = NULL, pill = FALSE, ...)

renderBadge(value, context = NULL, env = parent.frame(), quoted = FALSE)

Arguments

content
The text or label of the badge, defaults to NULL.

context
A character string specifying the visual context of the badge, one of "primary", "success", "info", "warning", or "danger", defaults to NULL.

pill
If TRUE the badge has rounder corners, defaults to FALSE.

...
Additional named argument passed on as HTML attributes to the parent element.

Examples

if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      listGroup(
        listItem(
          label = "Button clicks",
          badge = badge(
            id = "buttonClicks",
            content = 0
          )
        )
      ),
      button(
        id = "myButton",
        "Click here!"
      )
    ),
    server = function(input, output) {
      numClicks <- 0
      output$buttonClicks <- renderBadge({
        req(input$myButton)
        numClicks <<- numClicks + 1
        numClicks
      }, {
        if (numClicks > 19) {
          "danger"
        } else if (numClicks > 9) {
          "warning"
        } else {
          "info"
        }
      })
    }
  )
}