Buttons, button groups and toolbars

Button controls. buttonGroup and buttonToolbar are a means of grouping buttons aesthetically.

button(label = NULL, context = "secondary", outline = FALSE,
  block = FALSE, disabled = FALSE, textual = NULL, ...)

updateButton(id, count = NULL, context = NULL,
  session = getDefaultReactiveDomain())

Arguments

label
A character vector or tag elements to use as the button label, defaults to NULL.

context
Used to specify the visual context of the button, one of "primary", "secondary", "success", "info", "warning", "danger", or "link", defaults to "secondary".

Primary buttons are blue, secondary buttons are white and grey, buttons for success are green, informative buttons are a lighter blue, warning buttons are yellow, and buttons for danger are red. Specifying "link" makes the button render with the appearance of a link.

outline
If TRUE, the button's background is transparent, context is preserved, defaults to FALSE.

block
If TRUE, the button is block-level instead of inline, defaults to FALSE.

A block-level element will occupy the entire space of its parent element, thereby creating a "block."

disabled
If TRUE, the button renders in a disabled state, defaults to FALSE.

textual
Optional textual input, see text, to include as part of the button, defaults to NULL. If specified, the value of the button is the value of the text field.

...
Named arguments passed as HTML attributes to the button parent element.

Details

When adding a textual component to a button it is recommended to use placeholder text instead of a label.

See also

For more about block-level elements please refer to the block-level elements MDN reference section.

For more about input groups or buttons with text fields please refer to the button addons Bootstrap reference section.

Examples

button("Primary", context = "primary")

#> <div class="buckle-button input-group" data-count="0">
#>   <button type="button" class="btn btn-primary" role="button">Primary</button>
#> </div>

button("Secondary")

#> <div class="buckle-button input-group" data-count="0">
#>   <button type="button" class="btn btn-secondary" role="button">Secondary</button>
#> </div>

button("Success", context = "success")

#> <div class="buckle-button input-group" data-count="0">
#>   <button type="button" class="btn btn-success" role="button">Success</button>
#> </div>

button("Info", context = "info", outline = TRUE)

#> <div class="buckle-button input-group" data-count="0">
#>   <button type="button" class="btn btn-outline-info" role="button">Info</button>
#> </div>

button("\u2715", context = "warning")

#> <div class="buckle-button input-group" data-count="0">
#>   <button type="button" class="btn btn-warning" role="button">✕</button>
#> </div>

button("Danger!", context = "danger", disable = TRUE)

#> <div class="buckle-button input-group" data-count="0">
#>   <button type="button" class="btn btn-danger" disabled role="button">Danger!</button>
#> </div>

button(
  id = "search",
  "Go!",
  textual = inputs$text(
    placeholder = "search terms"
  )
)

#> <div class="buckle-button input-group" data-count="0" id="search">
#>   <span class="input-group-btn">
#>     <button type="button" class="btn btn-secondary" role="button">Go!</button>
#>   </span>
#>   <input class="form-control" type="text" placeholder="search terms"/>
#> </div>

button(
  "Check date",
  textual = inputs$date()
)

#> <div class="buckle-button input-group" data-count="0">
#>   <span class="input-group-btn">
#>     <button type="button" class="btn btn-secondary" role="button">Check date</button>
#>   </span>
#>   <input class="form-control" type="date"/>
#> </div>

if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      listGroup(
        listItem(
          tags$form(
            class = "form-inline",
          button(
            id = "simple",
            "Simple button"
          )
          ),
          badge = badge(id = "simpleClicks", 0)
        ),
        listItem(
          tags$form(
            class = "form-inline",
          button(
             id = "textual",
             label = "Textual button",
             textual = inputs$text(
               placeholder = "hello, world!"
             )
            )
          ),
          badge = badge(id = "textualClicks", 0)
        ),
        listItem(
          button(
            id = "reset",
            label = "Reset",
            context = "primary"
          )
        )
      )
    ),
    server = function(input, output) {
      output$simpleClicks <- renderBadge(input$simple$count)

      output$textualClicks <- renderBadge(
        value = {
          input$textual$count
        },
        context = {
          if (input$textual$count > 5) {
            "warning"
          } else {
            "default"
          }
        }
      )

      observeEvent(input$reset, ignoreInit = TRUE, {
        updateButton("simple", count = 0)
        updateButton("textual", count = 0)
      })
    }
  )
}