Cards

Handy card tools.

card(header = NULL, block = NULL, footer = NULL, text = "left",
  context = NULL, outline = FALSE, inverse = FALSE)

Arguments

header
Card header, a character vector or tag, defaults to NULL.

block
The card body, a character vector or tag(s), defaults to NULL.

footer
Card footer, a character vector or tag, defaults to NULL.

text
One of "left", "right", or "center", specifying the text alignment of the card, defaults to "left".

context
One of "primary", "success", "info", "warning", or "danger", used to add a visual context to the card, defaults to NULL.

outline
If TRUE, the card context is styled as the card outline rather than the card background, defaults to FALSE.

inverse
If TRUE, the card has a dark background and light text as opposed to the default dark text on light background, defaults to FALSE.