Grid System and Responsive Layout

Flexbox, grid layout, containers, rows, and columns (oh my). Arguments width and order can be applied depending on the width of the page, see the details section for more information.

col(..., width = NULL, offset = NULL, align = NULL, order = NULL)

row(..., align = NULL, justify = NULL, gutters = TRUE)

container(...)

Arguments

...
Child elements of the container, row, or col parent element or additional named arguments passed as HTML attributes or additional classes to the respective parent element.

width
1 through 12 or "auto", defaults to NULL, alternate widths for different viewports may be specified using a named multi-length vector or list, see details section.

If the width of multiple columns totals more than 12, columns are greedily wrapped into multiple rows.

If "auto", then the column fits to the width of the content.

If NULL, the column will fit to its parent container. When multiple columns have a NULL width, the parent element width is divided equally amongst all such columns. For example, two columns for which width is NULL will split their parent container's width equally in two.

offset
1 through 12, defaults to NULL, alternate offsets for different viewports may be specified using a named multi-length vector or list, see details section. Use offset to move columns to the right by increasing the left margin of the column.

align
Vertical alignment of a row's columns or a particular column, one of "start", "center", or "end", defaults to NULL in which case no vertical alignment is applied.

If "start", the column or children columns align to top of row.

If "center", the column or children columns align to middle of row.

If "end", the column or children columns align to bottom of row.

order
One of "first", "last", or "unordered", defaults to NULL. A column marked as "first" is visually rearranged and appears as the first column in its row. A column marked as "last" is visually rearranged and appears as the last column in its row.

justify
Horizontal alignment of row columns, one of "start", "end", "center", "between", or "around", defaults to NULL in which case no horizontal alignment is applied.

If "start", children columns align to left side of row.

If "end", children columns align to right side of row.

If "center", children columns align to center side of row.

If "between", children columns are spread out equally across the width of the row, columns on the far left and right are flush with the edge of the row.

If "around", children columns are spread equally across the width of the container, padding is added between columns and the edges of the parent row.

gutters
Defaults to TRUE, children columns receive horizontal padding for aesthetic spacing, to remove this padding specify gutters as FALSE.

Details

One can specify alternate column offsets or widths for different viewports by using the following naming convention. A specific offset or width will be applied when the viewport is at least the corresponding size, see below.

unnamed, applies to all viewports.

sm, at least 576 pixels wide.

md, at least 768 pixels wide.

lg, at least 992 pixels wide.

xl, at least 1200 pixels wide.

See examples for more on how to make use of this functionality.

See also

For more information on and examples of containers, rows, and the grid layout, please refer to the bootstrap grid system page.

Examples

# A column that,
#   - always has a width of 3
#   - is always offset by 3
col(
  width = 3,
  offset = 3
)

#> <div class="col col-3 offset-3"></div>

# A column that,
#   - when viewport is small width is 2 and no offset
#   - when viewport is large width is 4 and offset is 3
col(
  width = c(sm = 2, lg = 4),
  offset = c(sm = 0, lg = 3)
)

#> <div class="col col-sm-2 col-lg-4 offset-sm-0 offset-lg-3"></div>

# A column that,
#   - always fits to its content
#   - has no offset up to medium
#   - when the viewport is medium offset is 2
#   - when the viewport is large offset is 3
#   - when the viewport is extra large offset is 5
col(
  width = "auto",
  offset = c(md = 2, lg = 3, xl = 5)
)

#> <div class="col col-auto offset-md-2 offset-lg-3 offset-xl-5"></div>

# A column that,
#   - is 12 units wide (full width of parent) until medium viewport
#   - at medium viewport the column fits to the content
col(
  width = c(12, md = "auto")
)

#> <div class="col col-12 col-md-auto"></div>