List Styling and Unstyling

Rebooted builder functions for <ul>, <ol>, <dl>, <dt>, and <dd>. Ordered and unordered lists can now be unstyled, removing bullets and some margins. Description lists make use of the container system.

bs$ul(..., styled = TRUE, inline = FALSE)

bs$ol(..., styled = TRUE, inline = FALSE)

bs$dl(...)

bs$dt(..., width = NULL, truncate = FALSE)

bs$dd(..., width = NULL)

Arguments

styled
When FALSE the default list style is removed and left margins are removed, but this applies only to the direct children elements, defaults to TRUE.

inline
When TRUE, list bullets are removed and the list is rendered inline, defaults to FALSE.

width
Description lists can make use of the row, column grid system. 1 through 12, defaults to NULL, alternate widths for different viewports may be specified using a named multi-length vector or list, see details section.

truncate
When TRUE long description titles are shortened and a "..." is added, defaults to FALSE.

Details

One can specify alternate column offsets or widths for different viewports by using the following naming convention. A specific offset or width will be applied when the viewport is at least the corresponding size, see below.

unnamed, applies to all viewports.

sm, at least 576 pixels wide.

md, at least 768 pixels wide.

lg, at least 992 pixels wide.

xl, at least 1200 pixels wide.

See examples for more on how to make use of this functionality.

Examples