Page Navigation

Create a nav.

nav(content = NULL, align = "left", width = NULL, column = FALSE,
  tabs = FALSE, pills = FALSE)

Arguments

content
A list of links, where links are constructed using tags$a.

To render a link in a disabled state name the corresponding list element(s) disabled, see examples below.

When building a tabbed interface with nav a default tab may be specified by naming an element active, see examples below.

align
Alignment of the nav, one of "left", "right", "center", or "horizontal", defaults to "left".

width
Used to indicate the nav should fill the width of its parent element, one of "fill" or "justify", defaults to NULL. If "justify" all nav links have equal width, when "fill" nav links are only equally spaced.

column
If TRUE, the nav is rendered as a horizontal stack instead of vertically, defaults to FALSE.

tabs
If TRUE, the nav is rendered as a set of tabs for a tabbed interface, defaults to FALSE. Specifying TRUE will also add a data-toggle HTML attribute to the nav links allowing them to toggle tabs.

pills
If TRUE, the nav is rendered as pills, defaults to FALSE.

Details

To center a horizontal nav specify align = "horiztonal" and width = "fill" or width = "justify".

See also

For more information about the Bootstrap nav please follow this link.

Examples

nav(
  list(
    tags$a(href = "#home", "Home"),
    tags$a(href = "#about", "About")
  )
)

#> <ul class="nav">
#>   <li class="nav-item">
#>     <a href="#home" class="nav-link">Home</a>
#>   </li>
#>   <li class="nav-item">
#>     <a href="#about" class="nav-link">About</a>
#>   </li>
#> </ul>

# active link
nav(
  list(
    active = tags$a(href = "#home", "Home"),
    tags$a(href = "#about", "About")
  )
)

#> <ul class="nav">
#>   <li class="nav-item">
#>     <a href="#home" class="nav-link active">Home</a>
#>   </li>
#>   <li class="nav-item">
#>     <a href="#about" class="nav-link">About</a>
#>   </li>
#> </ul>

# active tab
nav(
  tabs = TRUE,
  list(
    active = tags$a(href = "#home", "Home"),
    tags$a(href = "#about", "About"),
    tags$a(href = "#blog", "Blog")
  )
)

#> <ul class="nav nav-tabs" role="tablist">
#>   <li class="nav-item">
#>     <a href="#home" class="nav-link active" data-toggle="tab" role="tab">Home</a>
#>   </li>
#>   <li class="nav-item">
#>     <a href="#about" class="nav-link" data-toggle="tab" role="tab">About</a>
#>   </li>
#>   <li class="nav-item">
#>     <a href="#blog" class="nav-link" data-toggle="tab" role="tab">Blog</a>
#>   </li>
#> </ul>