Progress bars

Create progress bars. A single progress bar may be compososed of multiple progress bar components. These components may have individual values, labels, and contexts. To create a composite progress bar pass multiple calls to bar to progress. Each bar component may have its own HTML id.

progress(...)

bar(value = 0, label = NULL, context = NULL, striped = FALSE,
  animated = FALSE, ...)

renderBar(value, label = NULL, env = parent.frame(), quoted = FALSE)

Arguments

...
bar elements passed to progress or named arguments passed as HTML attributes to the respective parent element.

value
An integer(s), between 0 and 100, specifying the initial value(s) of a progress bar or components of a progress bar, defaults to 0.

label
Label(s) for a progress bar or components of a progress bar, defaults to NULL.

context
Used to specify the visual context of the progress bar, "success", "info", "warning" or "danger".

striped
If TRUE, the progress bar has a striped gradient, defaults to FALSE.

animated
If TRUE, the progress bar is animated and striped is set to TRUE, defaults to FALSE.

Examples

if (interactive()) {
  library(shiny)

  shinyApp(
    ui = container(
      progress(
        bar(id = "clicks", value = 0)
      ),
      button(id = "inc", "inc progress")
    ),
    server = function(input, output, session) {
      output$clicks <- renderBar(
        min(input$inc$count, 10) / 10 * 100,
        min(input$inc$count, 10) / 10 * 100
      )
    }
  )
}