Radio, radio bars and groups

Create radio controls. Radio bars are stylized as button bars, but act as radio groups.

radio(content = NULL, value = NULL, name = NULL, inline = FALSE,
  disable = FALSE, ...)

radioBar(content = NULL, context = "secondary", name = NULL, ...)

Arguments

content
A character vector specifying the label of the radio control, defaults to NULL.

name
A character vector specifying the HTML name attribute of the input tag, defaults to NULL.

disable
If TRUE, the radio control is initialized with in a disabled state, defaults to FALSE.

...
Optional, additional named arguments passed as HTML attributes.

context
Used to specify the visual context of the radioBar, one of "primary", "secondary", "success", "info", "warning", "danger", or "link", defaults to "secondary".