Bootstrap Tables

New and improved tables, table headers, and table rows. Additional arguments have been added to help indicate and add shortcuts for possible Bootstrap stylizations.

bs$table(..., inverse = FALSE, striped = FALSE, borders = FALSE,
  hover = FALSE, compact = FALSE, scrollable = FALSE)

bs$thead(..., inverse = FALSE)

bs$tr(..., context = NULL, inverse = FALSE)

bs$th(..., context = NULL, inverse = FALSE)

bs$td(..., context = NULL, inverse = FALSE)

Arguments

context
One of "active", "primary", "success", "info", "warning", or "danger", applies contextual styling to table rows or cells, defaults to NULL.

Please be aware that "active" will not work when inverse is TRUE and "primary" only works when inverse is TRUE.

inverse
Table or table header renders with light text on a dark background, defaults to FALSE.

striped
Render table with zebra striped rows, defaults to FALSE.

borders
Render table with all borders, defaults to FALSE.

hover
Enable hover state on table rows, defaults to FALSE.

compact
Cut cell padding to make the table more compact, defaults to FALSE.

scrollable
If TRUE, the table has horizontal scroll on small viewports, defaults to FALSE.

Examples